// src/enums/ApiErrorMessage.ts
enum ApiErrorMessage {
    UNKNOWN_ERROR = 'UNKNOWN_ERROR',
    UNSPECIFIED_BODY = 'UNSPECIFIED_BODY',
    DOTNET_API_ERROR = 'DOTNET_API_ERROR',
    FIX_SYNTAX_ERRORS = 'FIX_SYNTAX_ERRORS',
    ONE_INTERFACE_LIMIT = 'ONE_INTERFACE_LIMIT',
    STATE_IDENTIFIER_IN_INTERFACE_REQUIRED = 'STATE_IDENTIFIER_IN_INTERFACE_REQUIRED',
    STATE_NAME_MUST_BE_CAPITALIZED = 'STATE_NAME_MUST_BE_CAPITALIZED',
    MAX_FIVE_PROPERTIES_ALLOWED_IN_STATE = 'MAX_FIVE_PROPERTIES_ALLOWED_IN_STATE',
    ONLY_PRIMITIVES_SUPPORTED_AS_STATE_PROPERTY_TYPES = 'ONLY_PRIMITIVES_SUPPORTED_AS_STATE_PROPERTY_TYPES'
}

export default ApiErrorMessage